Player movement update
Line 50 may not be needed, but as I understand it, normalizing movement is a standard for diagonal movement.

In order to fix the half move speed issue while grounded and pressing up and down, I'm adding half of the move speed back into the speed calculation as seen on line 52.

On a minor note, I updated the animation speed from 5FPS to 25.